+++ code/packages/store-azure-web-pubsub/src/index.ts
@@ -1,8 +1,0 @@
-// SPDX-FileCopyrightText: 2025 2025 INDUSTRIA DE DISEÑO TEXTIL S.A. (INDITEX S.A.)
-//
-// SPDX-License-Identifier: Apache-2.0
-
-export { WeaveStoreAzureWebPubsub } from "./store-azure-web-pubsub";
-export { WeaveStoreAzureWebPubSubSyncHost } from "./host";
-export * from "./constants";
-export * from "./types";