+++ code/CHANGELOG.md
@@ -13,6 +13,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.24.4] - 2025-04-14
+
+### Added
+
+- [#105](https://github.com/InditexTech/weavejs/issues/105) Fix store packages external depdendencies
+
+## [0.24.3] - 2025-04-14
+
+### Added
+
+- [#103](https://github.com/InditexTech/weavejs/issues/103) Fix externalized dependencies on store packages
+
+## [0.24.2] - 2025-04-14
+
+### Added
+
+- [#101](https://github.com/InditexTech/weavejs/issues/101) Fix missing `emittery` dependency package on the `@inditextech/weavejs-store-azure-web-pubsub` package
+
+## [0.24.1] - 2025-04-14
+
+### Added
+
+- [#99](https://github.com/InditexTech/weavejs/issues/99) Fix proxyPolicy is not supported in browser environment on `@inditextech/weavejs-store-azure-web-pubsub` package
+
+## [0.24.0] - 2025-04-14
+
+### Added
+
+- [#95](https://github.com/InditexTech/weavejs/issues/95) Refactor server part for websockets store
+- [#97](https://github.com/InditexTech/weavejs/issues/97) Refactor server part for Azure web pubsub store
+
 ## [0.23.0] - 2025-04-10
 
 ### Added
@@ -199,7 +230,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#8](https://github.com/InditexTech/weavejs/issues/8) ESLint OSS rules
 - [#2](https://github.com/InditexTech/weavejs/issues/2) Improve rendering handling
 
-[Unreleased]: https://github.com/InditexTech/weavejs/compare/0.23.0...HEAD
+[Unreleased]: https://github.com/InditexTech/weavejs/compare/0.24.4...HEAD
+
+[0.24.4]: https://github.com/InditexTech/weavejs/compare/0.24.3...0.24.4
+
+[0.24.3]: https://github.com/InditexTech/weavejs/compare/0.24.2...0.24.3
+
+[0.24.2]: https://github.com/InditexTech/weavejs/compare/0.24.1...0.24.2
+
+[0.24.1]: https://github.com/InditexTech/weavejs/compare/0.24.0...0.24.1
+
+[0.24.0]: https://github.com/InditexTech/weavejs/compare/0.23.0...0.24.0
 
 [0.23.0]: https://github.com/InditexTech/weavejs/compare/0.22.0...0.23.0